Add files via upload
🔹 Smoke Testing
This is done after API development is complete. Simply validate if the APIs are working and nothing breaks.

🔹 Functional Testing
This creates a test plan based on the functional requirements and compares the results with the expected results.

🔹 Integration Testing
This test combines several API calls to perform end-to-end tests. The intra-service communications and data transmissions are tested.

🔹 Regression Testing
This test ensures that bug fixes or new features shouldn’t break the existing behaviors of APIs.

🔹 Load Testing
This tests applications’ performance by simulating different loads. Then we can calculate the capacity of the application.

🔹 Stress Testing
We deliberately create high loads to the APIs and test if the APIs are able to function normally.

🔹 Security Testing
This tests the APIs against all possible external threats.

🔹 UI Testing
This tests the UI interactions with the APIs to make sure the data can be displayed properly.

🔹 Fuzz Testing
This injects invalid or unexpected input data into the API and tries to crash the API. In this way, it identifies the API vulnerabilities.